Standarize naming conventions for lib classes
