+++ app/controllers/application_controller.rb
@@ -1,4 +1,4 @@
-require 'services/auth_token'
+
 class ApplicationController < ActionController::API
   def authorized
     header = request.headers['Authorization']
@@ -6,7 +6,7 @@ class ApplicationController < ActionController::API
       header_token = header.split(' ')[-1]
 
       begin
-        user_id = DecodeTokenService.fire(header_token)
+        user_id = Services::DecodeTokenService.fire(header_token)
         @current_user = User.find(user_id)
       rescue ActiveRecord::RecordNotFound => e
         render json: { errors: e.message }, status: :unauthorized
